pyproject.toml
[tool.poetry]
name = "publisher"
description = "publisher"
version = "0.1.0"
readme = "README.md"
authors = ["myapp.com"]


[tool.poetry.dependencies]
python = "3.9.15"

django = { version = "^3.2", extras = ["argon2"] }
django-split-settings = "^1.2"
django-axes = "^5.41"
django-csp = "^3.7"
django-health-check = "^3.16"
django-http-referrer-policy = "^1.1"
django-permissions-policy = "^4.14"
django-stubs-ext = "^0.8"

psycopg2-binary = "^2.9"
gunicorn = "^20.0"
python-decouple = "^3.8"
structlog = "^22.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
celery = {extras = ["reddis"], version = "^5.2.7"}
redis = "^4.5.4"
flake8-annotations = "1.2.0"
pytest-faker = "^2.0.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^3.6"
django-querycount = "^0.8"
django-migration-linter = "^4.1"
django-extra-checks = "^0.13"
nplusone = "^1.0"

wemake-python-styleguide = "^0.17"
flake8-pytest-style = "^1.7"
flake8-django = "^1.1"
flake8-logging-format = "^0.9"
nitpick = "^0.33"
doc8 = "^1.0"

pytest = "^7.2"
pytest-django = "^4.5"
pytest-cov = "^4.0"
django-coverage-plugin = "^3.0"
pytest-randomly = "^3.12"
pytest-timeout = "^2.1"
django-test-migrations = "^1.2"
hypothesis = "^6.70"

mypy = "^1.1"
django-stubs = "^1.16"

yamllint = "^1.30"
safety = "^2.3"
dotenv-linter = "^0.4"
polint = "^0.4"
dennis = "^1.1"
dump-env = "^1.3"
ipython = "^8.12"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^6.1"
sphinx-autodoc-typehints = "^1.19"
tomli = "^2.0"


[build-system]
requires = ["poetry-core>=1.3.0"]
build-backend = "poetry.core.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/0.17.0/styles/nitpick-style-wemake.toml"
